cf/requirements.txt
Django==2.2.7
gunicorn==19.9.0
djangorestframework==3.10.3
whitenoise==4.1.4
